src/requirements.txt
amqp==5.0.7
appdirs==1.4.4
asgiref==3.4.1
attrs==20.3.0
beautifulsoup4==4.9.3
billiard==3.6.4.0
celery==5.2.1
certifi==2020.12.5
cffi==1.15.1
cfgv==3.2.0
chardet==4.0.0
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cloudant==2.14.0
coreapi==2.3.3
coreschema==0.0.4
cryptocode
distlib==0.3.1
dj-database-url==0.5.0
Django==3.2
django-bootstrap4==2.3.1
django-celery-results==2.2.0
django-environ==0.4.5
djangorestframework==3.12.2
drf-yasg==1.20.0
factory-boy==3.2.0
Faker==5.4.0
filelock==3.0.12
flake8==3.8.4
gevent==21.12.0
greenlet==1.1.2
gunicorn==20.0.4
identify==1.5.12
idna==2.10
importlib-metadata==2.1.1
inflection==0.5.1
iniconfig==1.1.1
itypes==1.2.0
Jinja2==2.11.2
kombu==5.2.2
MarkupSafe==1.1.1
mccabe==0.6.1
mysqlclient==2.1.1
nodeenv==1.5.0
packaging==20.8
parameterized==0.8.1
Pillow
pluggy==0.13.1
pre-commit==2.9.3
prompt-toolkit==3.0.24
psycopg2-binary
py==1.10.0
pycodestyle==2.6.0
pycparser==2.21
pycryptodomex==3.15.0
pyflakes==2.2.0
PyJWT==2.3.0
pyparsing==2.4.7
pytest==6.2.1
pytest-django==4.1.0
python-dateutil==2.8.1
pytz==2020.4
PyYAML==5.3.1
requests==2.25.1
ruamel.yaml
ruamel.yaml.clib
shortuuid==1.0.1
six==1.15.0
soupsieve==2.1
sqlparse==0.4.1
text-unidecode==1.3
toml==0.10.2
twilio==7.8.0
typing-extensions==3.7.4.3
uritemplate==3.0.1
urllib3==1.26.2
vine==5.0.0
virtualenv==20.3.0
wcwidth==0.2.5
zipp==3.4.0
zope.event==4.5.0
zope.interface==5.4.0
